feat: ReadStatusController 구현

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicReadStatusService.java
@@ -3,9 +3,7 @@ package com.sprint.mission.discodeit.service.basic;
 import com.sprint.mission.discodeit.common.ErrorMessages;
 import com.sprint.mission.discodeit.dto.request.create.ReadStatusCreateRequest;
 import com.sprint.mission.discodeit.dto.request.update.ReadStatusUpdateRequest;
-import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.ReadStatus;
-import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
@@ -13,6 +11,7 @@ import com.sprint.mission.discodeit.service.ReadStatusService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
 
@@ -25,48 +24,35 @@ public class BasicReadStatusService implements ReadStatusService {
 
     @Override
     public ReadStatus create(ReadStatusCreateRequest request) {
-        User user = userRepository.find(request.userId())
-                .orElseThrow(() -> new IllegalArgumentException(
-                        ErrorMessages.format("User", ErrorMessages.ERROR_NOT_FOUND)
-                ));
+        UUID userId = request.userId();
+        UUID channelId = request.channelId();
 
-        Channel channel = channelRepository.find(request.channelId())
-                .orElseThrow(() -> new IllegalArgumentException(
-                        ErrorMessages.format("Channel", ErrorMessages.ERROR_NOT_FOUND)
-                ));
-
-
-        //가장 최근에 읽은 시간 하나만 두기위해 중복 체크
-        boolean alreadyExists = readStatusRepository.findAllByUserId(request.userId()).stream()
-                .anyMatch(rs -> rs.getChannelId().equals(request.channelId()));
-
-        if (alreadyExists) {
-            throw new IllegalStateException(
-                    ErrorMessages.format("Message", ErrorMessages.ERROR_EXISTS)
-            );
+        if (!userRepository.existsById(userId)) {
+            throw new IllegalArgumentException(ErrorMessages.format("user", ErrorMessages.ERROR_NOT_FOUND));
+        }
+        if (!channelRepository.existsById(channelId)) {
+            throw new IllegalArgumentException(ErrorMessages.format("channel", ErrorMessages.ERROR_NOT_FOUND));
+        }
+        if (readStatusRepository.findAllByUserId(userId).stream()
+                .anyMatch(readStatus -> readStatus.getChannelId().equals(channelId))) {
+            throw new IllegalArgumentException(ErrorMessages.format("ReadStatus with user and channel", ErrorMessages.ERROR_EXISTS));
         }
 
-        ReadStatus readStatus = new ReadStatus(
-                request.userId(),
-                request.channelId(),
-                request.lastReadTime()
-        );
-
-        readStatusRepository.create(readStatus);
-        return readStatus;
+        Instant lastReadTime = request.lastReadTime();
+        ReadStatus readStatus = new ReadStatus(userId, channelId, lastReadTime);
+        return readStatusRepository.create(readStatus);
     }
 
     @Override
     public ReadStatus find(UUID id) {
-        ReadStatus readStatus = readStatusRepository.find(id).orElseThrow(()-> new IllegalArgumentException(
+        return readStatusRepository.find(id).orElseThrow(()-> new IllegalArgumentException(
                 ErrorMessages.format("ReadStatus", ErrorMessages.ERROR_NOT_FOUND)));
-
-        return readStatus;
     }
 
     @Override
     public List<ReadStatus> findAllByUserId(UUID userId) {
-        return readStatusRepository.findAllByUserId(userId);
+        return readStatusRepository.findAllByUserId(userId).stream()
+                .toList();
     }
 
     @Override
@@ -74,15 +60,16 @@ public class BasicReadStatusService implements ReadStatusService {
         ReadStatus readStatus = readStatusRepository.find(readStatusId).orElseThrow(()-> new IllegalArgumentException(
                 ErrorMessages.format("ReadStatus", ErrorMessages.ERROR_NOT_FOUND)));
 
-        readStatus.setLastReadTime(request.newReadTime());
+        readStatus.update(request.newReadTime());
         return readStatusRepository.create(readStatus);
     }
 
     @Override
     public void delete(UUID id) {
-        ReadStatus readStatus = readStatusRepository.find(id).orElseThrow(()-> new IllegalArgumentException(
-                ErrorMessages.format("ReadStatus", ErrorMessages.ERROR_NOT_FOUND)));
-
+        if (!readStatusRepository.existsById(id)) {
+            throw new IllegalArgumentException(
+                    ErrorMessages.format("ReadStatus", ErrorMessages.ERROR_NOT_FOUND));
+        };
         readStatusRepository.deleteById(id);
     }
 }
